Ensure code coverage analysis is provided.
<h1>Rationale:</h1>

This commit's raison d'etre is to ensure that both the sample app and the library provide code coverage statistics and a report for each.  The report clearly shows that we are starting form the grounded perch of 0% code coverage, i.e. no tests ensure that any lines of production code are executed, a great position to be in for a TDD developer.

<h1>File changes:</h1>

modified:   shelib/build.gradle

- Summary: ensure that the Sdk is at the current level; turn on code coverage analysis in the Android Gradle plugin.